LayoutTests/platform/ios-sim-deprecated/fast/canvas/webgl/oes-texture-half-float-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_ENUM: texImage2D: invalid texture type
This test verifies the functionality of OES_texture_half_float with null/non-null ArrayBufferView

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebGL context exists
Testing texture creation with extension disabled, format RGBA, and data null. Expect Failure
PASS getError was expected value: NO_ERROR : texture parameter setup should succeed
PASS getError was expected value: INVALID_ENUM : Half floating point texture must be disallowed if OES_texture_half_float isn't enabled
PASS No OES_texture_half_float support. This is legal

PASS successfullyParsed is true

TEST COMPLETE
